Restructure node app to enable integration tests; write first tests
