+++ package.json
@@ -66,6 +66,7 @@
     "react-dom": "^0.14.7",
     "sinon": "^1.17.2",
     "sinon-chai": "^2.8.0",
+    "supertest": "^1.1.0",
     "vinyl-source-stream": "^1.1.0",
     "webpack": "^1.12.9",
     "webpack-stream": "^3.1.0"
