+++ test/integration/api/transactions/index.js
@@ -1,5 +1,7 @@
 import pg from 'pg-promise';
+import request from 'supertest';
 import dbConfig from '../../../../server/api/util/db-config';
+import app from '../../../../server/app';
 
 const pgp = pg();
 
@@ -11,6 +13,18 @@ function getInsertQuery(values) {
     text: `INSERT INTO ${TABLE_NAME} (description, value, date) VALUES ($1, $2, $3)`,
     values: [values.description, values.value, values.date]
   };
+}
+
+function hasDataAttr(res) {
+  if (!('data' in res.body)) {
+    throw new Error('Missing `data` attribute.');
+  }
+}
+
+function dataIsArray(res) {
+  if (!Array.isArray(res.body.data)) {
+    return new Error('The `data` attribute is not an array.');
+  }
 }
 
 describe('Transactions', () => {
@@ -25,13 +39,32 @@ describe('Transactions', () => {
     return Promise.all(queries.map(q => db.none(q)));
   });
 
-  describe('A mock test', () => {
-    it('should pass', () => {
-      var db = pgp(dbConfig());
+  describe('/transactions', () => {
+    it('should return 200', done => {
+      request(app())
+        .get('/transactions')
+        .set('Accept', 'application/json')
+        .expect(200)
+        .end(function(err, res){
+          if (err) { return done(err); }
+          done();
+        });
+    });
 
-      return db.any(`SELECT * FROM ${TABLE_NAME}`)
-        .then(result => {
-          expect(result).to.have.length(2);
+    it('should have a data attribute that is an array of length 2', done => {
+      request(app())
+        .get('/transactions')
+        .set('Accept', 'application/json')
+        .expect(hasDataAttr)
+        .expect(dataIsArray)
+        .expect(res => {
+          if (res.body.data.length !== 2) {
+            return new Error('The length of the array was not 2.');
+          }
+        })
+        .end(function(err, res){
+          if (err) { return done(err); }
+          done();
         });
     });
   });
