+++ test/setup/node.js
@@ -1,3 +1,5 @@
+global.TESTING = true;
+
 global.chai = require('chai');
 global.sinon = require('sinon');
 global.chai.use(require('sinon-chai'));
